requirements.txt
Flask==0.11

# Application server for both development and production.
gunicorn==19.4.5

# Database and Workers
Flask-Mongoengine

# Extensions.
Flask-Script
Jsonschema
PyJWT
